Fix import cycle, move termination into switch case, use concurrentExecutor
- Move terminateAllAgentLoopWorkflowsForConversation to separate file
  (terminate.ts) to break import cycle via conversation_resource
- Use concurrentExecutor for parallel workflow termination
- Move termination + updateMetadata into each switch case branch
  instead of using a separate if after the switch
